Move Settings to its own crate
