+++ crates/go_to_line/Cargo.toml
@@ -11,5 +11,6 @@ doctest = false
 text = { path = "../text" }
 editor = { path = "../editor" }
 gpui = { path = "../gpui" }
+settings = { path = "../settings" }
 workspace = { path = "../workspace" }
 postage = { version = "0.4", features = ["futures-traits"] }
