+++ crates/project_panel/Cargo.toml
@@ -10,6 +10,7 @@ doctest = false
 [dependencies]
 gpui = { path = "../gpui" }
 project = { path = "../project" }
+settings = { path = "../settings" }
 theme = { path = "../theme" }
 util = { path = "../util" }
 workspace = { path = "../workspace" }
