+++ crates/search/src/buffer_search.rs
@@ -9,7 +9,8 @@ use gpui::{
 use language::OffsetRangeExt;
 use project::search::SearchQuery;
 use std::ops::Range;
-use workspace::{ItemHandle, Pane, Settings, ToolbarItemLocation, ToolbarItemView};
+use workspace::{ItemHandle, Pane, ToolbarItemLocation, ToolbarItemView};
+use settings::Settings;
 
 action!(Deploy, bool);
 action!(Dismiss);
