+++ crates/workspace/src/workspace.rs
@@ -2,7 +2,6 @@ pub mod lsp_status;
 pub mod menu;
 pub mod pane;
 pub mod pane_group;
-pub mod settings;
 pub mod sidebar;
 mod status_bar;
 mod toolbar;
@@ -31,7 +30,7 @@ pub use pane::*;
 pub use pane_group::*;
 use postage::prelude::Stream;
 use project::{fs, Fs, Project, ProjectEntryId, ProjectPath, Worktree};
-pub use settings::Settings;
+use settings::Settings;
 use sidebar::{Side, Sidebar, SidebarItemId, ToggleSidebarItem, ToggleSidebarItemFocus};
 use status_bar::StatusBar;
 pub use status_bar::StatusItemView;
